Update 230209
For some reason livewire refuses to validate both the select2 component and the wysiwyg editor refuses to put its content into the targeted textarea, so livewire also couldn't valdiate that (since it's empty).
I've done 3 days of research on the topic but found nothing. IDK what I'm doing wrong nor how to fix it.

This update mostly contains component updates, and new actions (Update and CreatePostAction).
I also created the validation method on the CreatePostForm component and commented out the rules it takes from PostFormValidation as a static constant.